+++ stockflux-securities-master/src/App.js
@@ -1,8 +1,8 @@
 import React from "react";
 import "stockflux-components";
 import Components from "stockflux-components";
-import SecuritiesTable from "./components/SecuritiesTable";
-import InputForm from "./components/InputForm";
+import SecuritiesTable from "./components/securities-table/SecuritiesTable";
+import InputForm from "./components/input-form/InputForm";
 import { Redirect, Route, Switch } from "react-router-dom";
 
 import "./App.css";
