+++ stockflux-securities-master/src/components/alert/Alert.js
