+++ stockflux-securities-master/src/components/confirmation/Confirmation.css
