+++ stockflux-securities-master/src/components/formControls/toggle-switch/ToggleSwitch.js
